requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
pandas==2.0.3
WeasyPrint==59.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.21
email-validator==2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
